Split files to prepare for pulling in new function
