+++ config/splat.us.dra.yaml
@@ -178,6 +178,7 @@ segments:
       - [0x73AAC, c, 73AAC]
       - [0x75DA0, c, 75DA0]
       - [0x75F54, c, 75F54]
+      - [0x7A4D0, c, 7A4D0]
       - [0x8D3E8, c, 8D3E8]
       - [0x953A0, c, 953A0]
       - [0x962AC, sbss]
